🔋👠 Checkpoint ./server.py:8394924/1018 ./views/login.html:8394924/2981 ./views/index.html:8394924/1464

+++ server.py
@@ -2,16 +2,17 @@
 # -*- coding: utf-8 -*-
 
 import os
-from flask import Flask, request, render_template, jsonify, send_file, url_for, redirect
+from flask import Flask, request, render_template, jsonify, send_file, url_for, redirect, session
 import uuid
 
 # Support for gomix's 'front-end' and 'back-end' UI.
 app = Flask(__name__, static_folder='public', template_folder='views')
 
 # Set the app secret key from the secret environment variables
-app.secret = os.environ.get('SENHA')
+app.secret_key = os.environ.get('SECRET')
 UPLOAD_FOLDER = 'uploads'
 app.config['UPLOAD_FOLDER'] = UPLOAD_FOLDER
+app.config['USERS'] = {os.environ.get('USER'): os.environ.get('SENHA')}
 
 #def generate_unique_filename(filename):
 #   ext = filename.rsplit('.', 1)[1]
@@ -21,6 +22,21 @@ app.config['UPLOAD_FOLDER'] = UPLOAD_FOLDER
 @app.route('/', methods=['GET'])
 def index():
     return render_template('index.html', qntd = len(os.listdir(app.config['UPLOAD_FOLDER'])), files=os.listdir(app.config['UPLOAD_FOLDER']))
+
+@app.route('/login', methods=['GET', 'POST'])
+def login():
+    if request.method == 'POST':
+        username = request.form['username']
+        password = request.form['password']
+        if username in app.config['USERS'] and app.config['USERS'][username] == password:
+            session['username'] = username
+            return redirect(url_for('index'))
+    return render_template('login.html')  
+
+@app.route('/logout')
+def logout():
+    session.pop('username', None)
+    return redirect(url_for('index'))
 
 @app.route('/upload', methods=['POST'])
 def upload_file():
@@ -45,10 +61,9 @@ def download_file(filename):
 @app.route('/remove/<filename>', methods=['POST'])
 def remove_file(filename):
     senha = request.form.get('senha')
-    if senha == app.secret:
-      if os.path.exists(os.path.join(app.config['UPLOAD_FOLDER'], filename)):
-          file_path = os.path.join(app.config['UPLOAD_FOLDER'], filename)
-          os.remove(file_path)
+    if 'username' in session and os.path.exists(os.path.join(app.config['UPLOAD_FOLDER'], filename)):
+        file_path = os.path.join(app.config['UPLOAD_FOLDER'], filename)
+        os.remove(file_path)
     return redirect(url_for('index'))
   
 if __name__ == '__main__':
